Centralized pollserver
This pollserver design is based on the Golang netpoll code.
Using a centralized pollserver allows us to block on golang
primitives instead of cgo calls for all calls that would block.
Also it decreases the number of open filedescriptors by a factor
of 3 (previously each socket opened 1 normal fd and 2 poll fd's).